% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{print.CompoML}
\alias{print.CompoML}
\title{Print the analysis results of the proportional means model}
\usage{
\method{print}{CompoML}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{\link{CompoML}}.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Print the results of \code{\link{CompoML}} object
}
\description{
Print the analysis results of the proportional means model.
}
\seealso{
\code{\link{CompoML}}, \code{\link{plot.CompoML}}.
}
\keyword{CompoML}
